+++ app/src/main/java/com/example/crumb/Models/Schedule.kt
@@ -10,8 +10,8 @@ data class Schedule(
     @ColumnInfo(name = "name") var name : String,
     @ColumnInfo(name = "date") var date : String,
     @ColumnInfo(name = "description") var description : String,
-    @ColumnInfo(name = "start_time") var start: Int,
-    @ColumnInfo(name = "end_time") var end : Int,
+    @ColumnInfo(name = "start_time") var start: Long,
+    @ColumnInfo(name = "end_time") var end : Long,
     @ColumnInfo(name = "duration") var duration : Int,
     @ColumnInfo(name = "steps") var steps : Int
 )
